backend/requirements.txt
# Python-зависимости backend'а.
#
# Зависимости добавляются пакетами по мере продвижения по IMPLEMENTATION_PLAN:
#   - 0.2: fastapi, uvicorn, pydantic, pydantic-settings, celery
#   - 0.3: sqlalchemy, alembic, asyncpg, psycopg (миграции + ORM)
#   - 1.1: python-jose, passlib, python-multipart (JWT auth)
#          + pytest, pytest-asyncio, httpx       (первые тесты)
#   - 2.x: numpy-financial                       (IRR/NPV расчёты)
#   - 5.x: openpyxl, python-pptx, WeasyPrint     (экспорт)

# --- Web framework ---
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.32.0,<0.33.0
pydantic>=2.9.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0

# --- Background jobs ---
celery[redis]>=5.4.0,<6.0.0

# --- Database / ORM / migrations ---
sqlalchemy[asyncio]>=2.0.36,<3.0.0
alembic>=1.14.0,<2.0.0
asyncpg>=0.30.0,<0.31.0
psycopg[binary]>=3.2.0,<4.0.0

# --- Auth (JWT + bcrypt) ---
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
# bcrypt 4.1+ удалил атрибут __about__.__version__, на котором завязан
# passlib 1.7.4 — приводит к ValueError "password cannot be longer than
# 72 bytes" в любом hash() (см. docs/ERRORS_AND_ISSUES.md). Pin к 4.0.x.
bcrypt>=4.0.0,<4.1.0
python-multipart>=0.0.9,<1.0.0

# --- Testing ---
pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0
httpx>=0.27.0,<1.0.0
